+++ packages/apollos-ui-connected/src/ContentCardConnected/index.js
@@ -1,6 +1,5 @@
 import ContentCardConnected from './ContentCardConnected';
 
-export contentCardComponentMapper from './contentCardComponentMapper';
 export GET_CONTENT_CARD from './getContentCard';
 
 export default ContentCardConnected;
