+++ packages/apollos-ui-connected/src/index.js
@@ -1,9 +1,5 @@
-export ContentCardConnected, {
-  contentCardComponentMapper,
-  GET_CONTENT_CARD,
-} from './ContentCardConnected';
+export ContentCardConnected, { GET_CONTENT_CARD } from './ContentCardConnected';
 export { fetchMoreResolver, share } from './utils';
-export HorizontalContentCardConnected from './HorizontalContentCardConnected';
 export {
   HorizontalLikedContentFeed,
   HorizontalLikedContentFeedConnected,
