--- packages/apollos-ui-kit/src/ThumbnailCard/Thumbnail.js
@@ -1,24 +1,0 @@
-import { StyleSheet } from 'react-native';
-import { compose } from 'recompose';
-
-import styled from '../styled';
-import ProgressiveImage from '../ProgressiveImage';
-import { getIsLoading } from '../isLoading';
-
-const Image = compose(
-  getIsLoading,
-  styled(
-    {
-      ...StyleSheet.absoluteFillObject,
-      height: '100%',
-      paddingTop: 0,
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    'Card.Image'
-  )
-)(ProgressiveImage);
-
-Image.propTypes = ProgressiveImage.propTypes;
-
-export default Image;
