+++ packages/apollos-ui-kit/src/content/FeaturedCard/FeaturedCard.stories.js
@@ -2,72 +2,39 @@ import React from 'react';
 import { storiesOf } from '@apollosproject/ui-storybook';
 import { ScrollView } from 'react-native';
 
-import BackgroundView from '../BackgroundView';
-import CenteredView from '../CenteredView';
-import { CardLabel } from '../Card';
+import BackgroundView from '../../BackgroundView';
+import CenteredView from '../../CenteredView';
+import { CardLabel } from '../../Card';
 
-import HighlightCard from '.';
+import FeaturedCard from '.';
 
-storiesOf('ui-kit/HighlightCard', module)
+storiesOf('ui-kit/FeaturedCard', module)
   .addDecorator((story) => (
     <BackgroundView>
       {/* eslint-disable-next-line react-native/no-inline-styles */}
       <CenteredView style={{ alignItems: 'stretch' }}>{story()}</CenteredView>
     </BackgroundView>
   ))
-  .add('examples', () => (
-    <ScrollView>
-      <HighlightCard
-        title={
-          'Are you telling me that you built a time machine out of a DeLorean?'
-        }
-        coverImage={[
-          {
-            uri: 'https://picsum.photos/800/1600/?random',
-          },
-        ]}
-        summary={
-          'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
-        }
-        hasAction
-        labelText={'Custom Label'}
-      />
-      <HighlightCard
-        title={
-          'Are you telling me that you built a time machine out of a DeLorean?'
-        }
-        coverImage={[
-          {
-            uri: 'https://picsum.photos/1600/800/?random',
-          },
-        ]}
-        summary={
-          'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
-        }
-        hasAction
-        isLiked={false}
-        labelText={'Custom Label'}
-      />
-      <HighlightCard
-        title={
-          'Are you telling me that you built a time machine out of a DeLorean?'
-        }
-        coverImage={[
-          {
-            uri: 'https://picsum.photos/800/800/?random',
-          },
-        ]}
-        summary={
-          'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
-        }
-        hasAction
-        isLiked
-        labelText={'Custom Label'}
-      />
-    </ScrollView>
+  .add('example', () => (
+    <FeaturedCard
+      title={
+        'Are you telling me that you built a time machine out of a DeLorean?'
+      }
+      coverImage={[
+        {
+          uri: 'https://picsum.photos/800/1600/?random',
+        },
+      ]}
+      summary={
+        'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
+      }
+      hasAction
+      isLive
+      isLiked
+    />
   ))
   .add('default', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -79,7 +46,7 @@ storiesOf('ui-kit/HighlightCard', module)
     />
   ))
   .add('actionIcon', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -93,7 +60,7 @@ storiesOf('ui-kit/HighlightCard', module)
     />
   ))
   .add('summary', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -108,7 +75,7 @@ storiesOf('ui-kit/HighlightCard', module)
     />
   ))
   .add('hasAction', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -121,7 +88,7 @@ storiesOf('ui-kit/HighlightCard', module)
     />
   ))
   .add('isLiked/false', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -134,7 +101,7 @@ storiesOf('ui-kit/HighlightCard', module)
     />
   ))
   .add('isLiked/true', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -146,8 +113,21 @@ storiesOf('ui-kit/HighlightCard', module)
       isLiked
     />
   ))
+  .add('isLive', () => (
+    <FeaturedCard
+      title={
+        'Are you telling me that you built a time machine out of a DeLorean?'
+      }
+      coverImage={[
+        {
+          uri: 'https://picsum.photos/800/1600/?random',
+        },
+      ]}
+      isLive
+    />
+  ))
   .add('isLoading', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -160,12 +140,12 @@ storiesOf('ui-kit/HighlightCard', module)
         'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
       }
       hasAction
-      labelText={'Custom Label'}
+      isLive
       isLoading
     />
   ))
   .add('LabelComponent', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -178,7 +158,7 @@ storiesOf('ui-kit/HighlightCard', module)
     />
   ))
   .add('labelText', () => (
-    <HighlightCard
+    <FeaturedCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -191,25 +171,25 @@ storiesOf('ui-kit/HighlightCard', module)
     />
   ))
   .add('theme', () => (
-    <HighlightCard
-      title={
-        'Are you telling me that you built a time machine out of a DeLorean?'
-      }
-      coverImage={[
-        {
-          uri: 'https://picsum.photos/800/1600/?random',
-        },
-      ]}
-      theme={{
-        colors: {
-          primary: 'salmon',
-        },
-      }}
-    />
-  ))
-  .add('theme', () => (
     <ScrollView>
-      <HighlightCard
+      <FeaturedCard
+        title={
+          'Are you telling me that you built a time machine out of a DeLorean?'
+        }
+        coverImage={[
+          {
+            uri: 'https://picsum.photos/800/1600/?random',
+          },
+        ]}
+        hasAction
+        theme={{
+          colors: {
+            primary: 'salmon',
+            secondary: 'salmon',
+          },
+        }}
+      />
+      <FeaturedCard
         title={
           'Are you telling me that you built a time machine out of a DeLorean?'
         }
@@ -219,12 +199,10 @@ storiesOf('ui-kit/HighlightCard', module)
           },
         ]}
         theme={{
-          colors: {
-            primary: 'salmon',
-          },
+          type: 'light',
         }}
       />
-      <HighlightCard
+      <FeaturedCard
         title={
           'Are you telling me that you built a time machine out of a DeLorean?'
         }
@@ -233,14 +211,17 @@ storiesOf('ui-kit/HighlightCard', module)
             uri: 'https://picsum.photos/800/1600/?random',
           },
         ]}
-        labelText={'Quote'}
-        hasAction
+        summary={
+          'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
+        }
+        isLive
         isLiked
+        hasAction
         theme={{
           type: 'light',
           colors: {
             primary: 'yellow',
-            white: 'dodgerblue',
+            secondary: 'dodgerblue',
             text: {
               primary: 'dodgerblue',
             },
