+++ packages/apollos-ui-kit/src/content/FeaturedCard/FeaturedCard.tests.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Providers from '../Providers';
-import { CardLabel } from '../Card';
+import Providers from '../../Providers';
+import { CardLabel } from '../../Card';
 
-import HighlightCard from '.';
+import FeaturedCard from '.';
 
-describe('HighlightCard', () => {
-  it('should render in a "tall" aspect ratio (maxAspectRatio)', () => {
+describe('FeaturedCard', () => {
+  it('should render', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -24,27 +24,10 @@ describe('HighlightCard', () => {
     );
     expect(tree).toMatchSnapshot();
   });
-  it('should render in a "wide" aspect ratio', () => {
-    const tree = renderer.create(
-      <Providers>
-        <HighlightCard
-          title={
-            'Are you telling me that you built a time machine out of a DeLorean?'
-          }
-          coverImage={[
-            {
-              uri: 'https://picsum.photos/1600/800/?random',
-            },
-          ]}
-        />
-      </Providers>
-    );
-    expect(tree).toMatchSnapshot();
-  });
   it('should render with a custom actionIcon', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -63,7 +46,7 @@ describe('HighlightCard', () => {
   it('should render with a summary', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -83,7 +66,7 @@ describe('HighlightCard', () => {
   it('should should render with an action "button"', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -101,7 +84,7 @@ describe('HighlightCard', () => {
   it('should should render as isLiked', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -119,7 +102,7 @@ describe('HighlightCard', () => {
   it('should should render a like icon', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -137,7 +120,7 @@ describe('HighlightCard', () => {
   it('should should render as isLive', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -155,7 +138,7 @@ describe('HighlightCard', () => {
   it('should render a loading state with isLoading', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -178,7 +161,7 @@ describe('HighlightCard', () => {
   it('should render with custom LabelComponent', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -196,7 +179,7 @@ describe('HighlightCard', () => {
   it('should render with custom labelText', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -214,7 +197,7 @@ describe('HighlightCard', () => {
   it('should render with a custom theme', () => {
     const tree = renderer.create(
       <Providers>
-        <HighlightCard
+        <FeaturedCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
