+++ packages/apollos-ui-kit/src/content/HighlightCard/HighlightCard.stories.js
@@ -2,39 +2,72 @@ import React from 'react';
 import { storiesOf } from '@apollosproject/ui-storybook';
 import { ScrollView } from 'react-native';
 
-import BackgroundView from '../BackgroundView';
-import CenteredView from '../CenteredView';
-import { CardLabel } from '../Card';
+import BackgroundView from '../../BackgroundView';
+import CenteredView from '../../CenteredView';
+import { CardLabel } from '../../Card';
 
-import FeaturedCard from '.';
+import HighlightCard from '.';
 
-storiesOf('ui-kit/FeaturedCard', module)
+storiesOf('ui-kit/HighlightCard', module)
   .addDecorator((story) => (
     <BackgroundView>
       {/* eslint-disable-next-line react-native/no-inline-styles */}
       <CenteredView style={{ alignItems: 'stretch' }}>{story()}</CenteredView>
     </BackgroundView>
   ))
-  .add('example', () => (
-    <FeaturedCard
-      title={
-        'Are you telling me that you built a time machine out of a DeLorean?'
-      }
-      coverImage={[
-        {
-          uri: 'https://picsum.photos/800/1600/?random',
-        },
-      ]}
-      summary={
-        'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
-      }
-      hasAction
-      isLive
-      isLiked
-    />
+  .add('examples', () => (
+    <ScrollView>
+      <HighlightCard
+        title={
+          'Are you telling me that you built a time machine out of a DeLorean?'
+        }
+        coverImage={[
+          {
+            uri: 'https://picsum.photos/800/1600/?random',
+          },
+        ]}
+        summary={
+          'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
+        }
+        hasAction
+        labelText={'Custom Label'}
+      />
+      <HighlightCard
+        title={
+          'Are you telling me that you built a time machine out of a DeLorean?'
+        }
+        coverImage={[
+          {
+            uri: 'https://picsum.photos/1600/800/?random',
+          },
+        ]}
+        summary={
+          'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
+        }
+        hasAction
+        isLiked={false}
+        labelText={'Custom Label'}
+      />
+      <HighlightCard
+        title={
+          'Are you telling me that you built a time machine out of a DeLorean?'
+        }
+        coverImage={[
+          {
+            uri: 'https://picsum.photos/800/800/?random',
+          },
+        ]}
+        summary={
+          'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
+        }
+        hasAction
+        isLiked
+        labelText={'Custom Label'}
+      />
+    </ScrollView>
   ))
   .add('default', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -46,7 +79,7 @@ storiesOf('ui-kit/FeaturedCard', module)
     />
   ))
   .add('actionIcon', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -60,7 +93,7 @@ storiesOf('ui-kit/FeaturedCard', module)
     />
   ))
   .add('summary', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -75,7 +108,7 @@ storiesOf('ui-kit/FeaturedCard', module)
     />
   ))
   .add('hasAction', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -88,7 +121,7 @@ storiesOf('ui-kit/FeaturedCard', module)
     />
   ))
   .add('isLiked/false', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -101,7 +134,7 @@ storiesOf('ui-kit/FeaturedCard', module)
     />
   ))
   .add('isLiked/true', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -113,21 +146,8 @@ storiesOf('ui-kit/FeaturedCard', module)
       isLiked
     />
   ))
-  .add('isLive', () => (
-    <FeaturedCard
-      title={
-        'Are you telling me that you built a time machine out of a DeLorean?'
-      }
-      coverImage={[
-        {
-          uri: 'https://picsum.photos/800/1600/?random',
-        },
-      ]}
-      isLive
-    />
-  ))
   .add('isLoading', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -140,12 +160,12 @@ storiesOf('ui-kit/FeaturedCard', module)
         'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
       }
       hasAction
-      isLive
+      labelText={'Custom Label'}
       isLoading
     />
   ))
   .add('LabelComponent', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -158,7 +178,7 @@ storiesOf('ui-kit/FeaturedCard', module)
     />
   ))
   .add('labelText', () => (
-    <FeaturedCard
+    <HighlightCard
       title={
         'Are you telling me that you built a time machine out of a DeLorean?'
       }
@@ -171,8 +191,25 @@ storiesOf('ui-kit/FeaturedCard', module)
     />
   ))
   .add('theme', () => (
+    <HighlightCard
+      title={
+        'Are you telling me that you built a time machine out of a DeLorean?'
+      }
+      coverImage={[
+        {
+          uri: 'https://picsum.photos/800/1600/?random',
+        },
+      ]}
+      theme={{
+        colors: {
+          primary: 'salmon',
+        },
+      }}
+    />
+  ))
+  .add('theme', () => (
     <ScrollView>
-      <FeaturedCard
+      <HighlightCard
         title={
           'Are you telling me that you built a time machine out of a DeLorean?'
         }
@@ -181,15 +218,13 @@ storiesOf('ui-kit/FeaturedCard', module)
             uri: 'https://picsum.photos/800/1600/?random',
           },
         ]}
-        hasAction
         theme={{
           colors: {
             primary: 'salmon',
-            secondary: 'salmon',
           },
         }}
       />
-      <FeaturedCard
+      <HighlightCard
         title={
           'Are you telling me that you built a time machine out of a DeLorean?'
         }
@@ -198,30 +233,14 @@ storiesOf('ui-kit/FeaturedCard', module)
             uri: 'https://picsum.photos/800/1600/?random',
           },
         ]}
-        theme={{
-          type: 'light',
-        }}
-      />
-      <FeaturedCard
-        title={
-          'Are you telling me that you built a time machine out of a DeLorean?'
-        }
-        coverImage={[
-          {
-            uri: 'https://picsum.photos/800/1600/?random',
-          },
-        ]}
-        summary={
-          'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
-        }
-        isLive
-        isLiked
+        labelText={'Quote'}
         hasAction
+        isLiked
         theme={{
           type: 'light',
           colors: {
             primary: 'yellow',
-            secondary: 'dodgerblue',
+            white: 'dodgerblue',
             text: {
               primary: 'dodgerblue',
             },
