+++ packages/apollos-ui-kit/src/content/HighlightCard/HighlightCard.tests.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Providers from '../Providers';
-import { CardLabel } from '../Card';
+import Providers from '../../Providers';
+import { CardLabel } from '../../Card';
 
-import DefaultCard from '.';
+import HighlightCard from '.';
 
-describe('DefaultCard', () => {
-  it('should render at the minAspectRatio bound (tall-ish)', () => {
+describe('HighlightCard', () => {
+  it('should render in a "tall" aspect ratio (maxAspectRatio)', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -24,16 +24,16 @@ describe('DefaultCard', () => {
     );
     expect(tree).toMatchSnapshot();
   });
-  it('should render at the images natural aspect ratio (between min/max)', () => {
+  it('should render in a "wide" aspect ratio', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
           coverImage={[
             {
-              uri: 'https://picsum.photos/1400/800/?random',
+              uri: 'https://picsum.photos/1600/800/?random',
             },
           ]}
         />
@@ -41,19 +41,20 @@ describe('DefaultCard', () => {
     );
     expect(tree).toMatchSnapshot();
   });
-
-  it('should render at the maxAspectRatio bound (wider than tall)', () => {
+  it('should render with a custom actionIcon', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
           coverImage={[
             {
-              uri: 'https://picsum.photos/2000/800/?random',
+              uri: 'https://picsum.photos/800/1600/?random',
             },
           ]}
+          actionIcon={'umbrella'}
+          hasAction
         />
       </Providers>
     );
@@ -62,7 +63,7 @@ describe('DefaultCard', () => {
   it('should render with a summary', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -79,10 +80,10 @@ describe('DefaultCard', () => {
     );
     expect(tree).toMatchSnapshot();
   });
-  it('should should render a like icon', () => {
+  it('should should render with an action "button"', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -91,7 +92,7 @@ describe('DefaultCard', () => {
               uri: 'https://picsum.photos/800/1600/?random',
             },
           ]}
-          isLiked={false}
+          hasAction
         />
       </Providers>
     );
@@ -100,7 +101,7 @@ describe('DefaultCard', () => {
   it('should should render as isLiked', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -115,10 +116,46 @@ describe('DefaultCard', () => {
     );
     expect(tree).toMatchSnapshot();
   });
+  it('should should render a like icon', () => {
+    const tree = renderer.create(
+      <Providers>
+        <HighlightCard
+          title={
+            'Are you telling me that you built a time machine out of a DeLorean?'
+          }
+          coverImage={[
+            {
+              uri: 'https://picsum.photos/800/1600/?random',
+            },
+          ]}
+          isLiked={false}
+        />
+      </Providers>
+    );
+    expect(tree).toMatchSnapshot();
+  });
+  it('should should render as isLive', () => {
+    const tree = renderer.create(
+      <Providers>
+        <HighlightCard
+          title={
+            'Are you telling me that you built a time machine out of a DeLorean?'
+          }
+          coverImage={[
+            {
+              uri: 'https://picsum.photos/800/1600/?random',
+            },
+          ]}
+          isLive
+        />
+      </Providers>
+    );
+    expect(tree).toMatchSnapshot();
+  });
   it('should render a loading state with isLoading', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -130,6 +167,8 @@ describe('DefaultCard', () => {
           summary={
             'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
           }
+          hasAction
+          isLive
           isLoading
         />
       </Providers>
@@ -139,7 +178,7 @@ describe('DefaultCard', () => {
   it('should render with custom LabelComponent', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -148,9 +187,7 @@ describe('DefaultCard', () => {
               uri: 'https://picsum.photos/800/1600/?random',
             },
           ]}
-          LabelComponent={
-            <CardLabel title={'Custom LabelComponent'} type={'primary'} />
-          }
+          LabelComponent={<CardLabel title={'Custom LabelComponent'} />}
         />
       </Providers>
     );
@@ -159,7 +196,7 @@ describe('DefaultCard', () => {
   it('should render with custom labelText', () => {
     const tree = renderer.create(
       <Providers>
-        <DefaultCard
+        <HighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -168,7 +205,29 @@ describe('DefaultCard', () => {
               uri: 'https://picsum.photos/800/1600/?random',
             },
           ]}
-          labelText={'Custom label text'}
+          labelText={'Quote'}
+        />
+      </Providers>
+    );
+    expect(tree).toMatchSnapshot();
+  });
+  it('should render with a custom theme', () => {
+    const tree = renderer.create(
+      <Providers>
+        <HighlightCard
+          title={
+            'Are you telling me that you built a time machine out of a DeLorean?'
+          }
+          coverImage={[
+            {
+              uri: 'https://picsum.photos/800/1600/?random',
+            },
+          ]}
+          theme={{
+            colors: {
+              primary: 'salmon',
+            },
+          }}
         />
       </Providers>
     );
