+++ packages/apollos-ui-kit/src/content/HorizontalDefaultCard/HorizontalDefaultCard.tests.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Providers from '../Providers';
+import Providers from '../../Providers';
 
 import HorizontalDefaultCard from '.';
