+++ packages/apollos-ui-kit/src/content/HorizontalHighlightCard/HorizontalHighlightCard.tests.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Providers from '../Providers';
-import { CardLabel } from '../Card';
+import Providers from '../../Providers';
+import { CardLabel } from '../../Card';
 
-import FeaturedCard from '.';
+import HorizontalHighlightCard from '.';
 
-describe('FeaturedCard', () => {
+describe('HorizontalHighlightCard', () => {
   it('should render', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -27,7 +27,7 @@ describe('FeaturedCard', () => {
   it('should render with a custom actionIcon', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -43,30 +43,10 @@ describe('FeaturedCard', () => {
     );
     expect(tree).toMatchSnapshot();
   });
-  it('should render with a summary', () => {
+  it('should render with an action "button"', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
-          title={
-            'Are you telling me that you built a time machine out of a DeLorean?'
-          }
-          coverImage={[
-            {
-              uri: 'https://picsum.photos/800/1600/?random',
-            },
-          ]}
-          summary={
-            'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
-          }
-        />
-      </Providers>
-    );
-    expect(tree).toMatchSnapshot();
-  });
-  it('should should render with an action "button"', () => {
-    const tree = renderer.create(
-      <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -81,10 +61,10 @@ describe('FeaturedCard', () => {
     );
     expect(tree).toMatchSnapshot();
   });
-  it('should should render as isLiked', () => {
+  it('should should render as disabled', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -93,16 +73,16 @@ describe('FeaturedCard', () => {
               uri: 'https://picsum.photos/800/1600/?random',
             },
           ]}
-          isLiked
+          disabled
         />
       </Providers>
     );
     expect(tree).toMatchSnapshot();
   });
-  it('should should render a like icon', () => {
+  it('should should render as a like icon', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -117,10 +97,10 @@ describe('FeaturedCard', () => {
     );
     expect(tree).toMatchSnapshot();
   });
-  it('should should render as isLive', () => {
+  it('should should render as isLiked', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -129,7 +109,7 @@ describe('FeaturedCard', () => {
               uri: 'https://picsum.photos/800/1600/?random',
             },
           ]}
-          isLive
+          isLiked
         />
       </Providers>
     );
@@ -138,7 +118,7 @@ describe('FeaturedCard', () => {
   it('should render a loading state with isLoading', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -147,12 +127,9 @@ describe('FeaturedCard', () => {
               uri: 'https://picsum.photos/800/1600/?random',
             },
           ]}
-          summary={
-            'The way I see it, if you’re going to build a time machine into a car, why not do it with some style?'
-          }
           hasAction
-          isLive
           isLoading
+          labelText={'Loading!'}
         />
       </Providers>
     );
@@ -161,7 +138,7 @@ describe('FeaturedCard', () => {
   it('should render with custom LabelComponent', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -179,7 +156,7 @@ describe('FeaturedCard', () => {
   it('should render with custom labelText', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
@@ -197,7 +174,7 @@ describe('FeaturedCard', () => {
   it('should render with a custom theme', () => {
     const tree = renderer.create(
       <Providers>
-        <FeaturedCard
+        <HorizontalHighlightCard
           title={
             'Are you telling me that you built a time machine out of a DeLorean?'
           }
