+++ packages/apollos-ui-kit/src/content/index.js
@@ -1,0 +1,8 @@
+export ContentCard from './ContentCard';
+// deprecate the use of these directly
+// should be <ContentCard type={"highlight"} horizontal />
+export DefaultCard from './DefaultCard';
+export HighlightCard from './HighlightCard';
+export FeaturedCard from './FeaturedCard';
+export HorizontalDefaultCard from './HorizontalDefaultCard';
+export HorizontalHighlightCard from './HorizontalHighlightCard';
